Make `task_id` argument in `task_status` optional
For several robots, the `task_status()` function will return the status
of the current executing task, and there is no need for the task id.

The `task_id` argument is kept as optional for now, but it should be
considered to remove the argument and leave `task_status()` to only
return the current task id.