Add `Task` as part of the mission hierarchy
The Mission has been restructured to consist of a set of Tasks which
again may consist of a set of Steps. What is now known as a Step
corresponds to the old concept of a Task (e.g. DriveToPose, TakeImage).

This enables tracking dependencies between steps through the concept of
a Task. If a MotionStep in a Task has failed the entire Task should be
considered as failed. If an InspectionStep has failed the Task will
continue but mark the Task as a whole as PartiallySuccessful. If all
InspectionSteps have failed the Task will be marked as failed.