Add modularized task selector interface
Implements a modular interface for task selection per mission.

Replaces the iterator in the Mission class responsible for selecting
the next task of the current mission.

When there are no more tasks to select for the current mission, the
TaskSelectorStop exception should be raised in order to comply with the
interface.