Fix fastapi endpoints returning correct type
Some of the endpoints returned an error message as string.
Now, the appropriate HTTPException with error message is raised instead.

Includes some refactoring.

Moved logic from scheduling controller to scheduling utilites to
improve readability and remove duplication of code.

Moved mission validator to scheduling utilities.

The tests for the scheduler router is rewritten for readability.
Missing tests for some of the endpoints are added.